+++ packages/utils/src/apis/admin/index.ts
@@ -2,4 +2,4 @@ export * from './faqApi';
 export * from './loginApi';
 export * from './imageApi';
 export * from './recruitApi';
-export * from './manageUserApi';
+export * from './adminManageUserApi';
